revert test_fleet_base
