+++ tools/parallel_UT_rule.py
@@ -942,7 +942,6 @@ FOURTH_HIGH_PARALLEL_JOB_NEW = [
     'test_sign_op',
     'selected_rows_functor_gpu_test',
     'test_fleet_base',
-    'test_fleet_base_deprecated',
     'test_logsumexp',
     'test_detection',
     'test_image_classification_fp16',
@@ -2232,7 +2231,6 @@ TETRAD_PARALLEL_JOB = [
     'test_analyzer_bert',
     'test_analyzer_googlenet',
     'test_fleet_base',
-    'test_fleet_base_deprecated',
     'test_dgc_momentum_op',
     'test_memcpy_op',
     'test_dgc_op',
